fix(ios): correct three defects found reviewing the unbuilt Swift
No Swift compiler was available here and download.swift.org is blocked by the
environment's network policy, so none of the Swift has ever been built. A
reading pass found three real defects, which is worth recording as evidence of
what review catches when a compiler cannot.

1. The coordinator was registered as a script message handler and never
   removed. WKUserContentController retains its handlers strongly, so pushing
   and popping editor screens would have accumulated a webview apiece — on a
   phone, a leak that ends in a jetsam kill rather than a visible bug. Now torn
   down in dismantleUIView, with the handler name declared once so registration
   and teardown cannot drift.

2. The navigation policy said one thing and did another. Its comment claimed
   the editor "must never navigate", while the code allowed everything that was
   not a tapped link — including any redirect a document could induce. It is now
   an allowlist of exactly one URL, the editor's own, with link taps reported to
   the host and everything else cancelled. An allowlist because a page is user
   content, and enumerating the ways content might try to navigate is a losing
   game.

3. The protocol version crossed the bridge as a Double and was compared for
   floating-point equality. The protocol declaration gains an `integer` field
   type — distinct from `number` so Swift gets an Int — and the version uses it.
   Non-integers are rejected by the webview's validator too, since a fractional
   value means the two ends disagree about the type rather than the value.

03-m0-spike.md gains a section saying plainly that the Swift is unproven, and
what to expect on the first build: strict-concurrency annotations are likely
needed on the coordinator, and were deliberately not added speculatively —
guessing at actor isolation without a compiler produces annotations that are
confidently wrong.

+++ apps/ios/PikosEditorBridge/Sources/PikosEditorBridge/EditorBridgeProtocol.swift
@@ -147,9 +147,9 @@ public enum EditorBridge {
     /// the webview was built against so the host can refuse a mismatch loudly
     /// rather than misbehave quietly.
     public struct IncomingReadyPayload: Codable, Equatable {
-        public let protocolVersion: Double
+        public let protocolVersion: Int
 
-        public init(protocolVersion: Double) {
+        public init(protocolVersion: Int) {
             self.protocolVersion = protocolVersion
         }
     }
